Add green deal sub header
Affected outcome:
outcome_bills_and_measures_on_benefits_not_eco_eligible.govspeak.erb

If the user is eligible for either a winter fuel payment or a cold
weather payment (i.e. help from the government), the section about the
green deal displays underneath the section about help from the
government.

In this circumstance a sub-heading should be added so that the user
knows that the help that follows can be applied for via the green deal
but not from a scheme funded by the government.

Regenerate test artefacts:

The green deal header text renders as expected if the user is eligible
for a winter fuel or cold weather payment.