Remove green deal message text
Affected outcome:
outcome_bills_and_measures_on_benefits_not_eco_eligible.govspeak.erb

This text is in the wrong place (between the sections on “boilers and
insulation” and “heating”) and should be removed. It has already been
added to the correct location in a previous commit.

Regenerate test artefacts:

The green deal text between the “boilers and insulation” and “heating”
sections is removed as expected.