Move reduced maintenance loan information
The reduced maintenance loan is not available via NHS as the other forms of
funding are, so it is not appropriate to include it in the list.

The information has been removed from the partial and added the to the lead-in
text instead.

Affects all uk medical, dental and healthcare students.