Update artefacts am-i-getting-minimum-wage
This is a preparatory commit where we regenerate all artefacts to match the
start date for the new rates.

The default date that the regression tests use as “today” to create the test
artefacts are set in [`data/configuration.yml`][2]. If no date is set for a
specific calculator, the default is used. We added a new date for
am-i-getting-minimum-wage (2016-10-01) because this is the start date of the new
date rates.

Expect to see changes in the artefacts because the [amount owed][1] has
increased in the past payment and current payment routes because of the extra
time that has passed in the tests.

[1]: https://github.com/alphagov/smart-answers/blob/c265170b8c94984daa83924bc50b3ec796b2b70d/lib/smart_answer/calculators/minimum_wage_calculator.rb#L121-L123
[2]: https://github.com/alphagov/smart-answers/blob/d4dd9e719aea5913dfad94bea8cd90720f48ea20/test/data/configurations.yml