Add function to retrieve the reference date for calculations
Add a function that returns today’s date or the date from an environment 
variable.

We have created this function for minimum-wage-calculator to support the
use of the RATES_QUERY_DATE environment variable. This variable was used
in RatesQuery, but we also need to access to this value from the 
calculator itself, so we have decided to expose this functionality in 
its own helper function.