Extend flow DSL with `external_related_links`
Panopticon is currently storing external links for smart answers.

Panopticon is going to be retired and we are going to store them into 
the content-store[1], via the publishing-api[2].
This commit extends the flow DSL to incorporate into the flow definition
the list of external links that could be displayed linked to the smart-answer.

This is a TEMPORARY UPDATE because this information will eventually be set 
via a publishing tool that will use content-store.

As a side note, the method `Flow#external_related_links` is used at the
same time as a `getter` and a `setter`, for this reason its implementation
feels a bit out of Ruby usual coding practices.

[1]: https://github.com/alphagov/content-store
[2]: https://github.com/alphagov/publishing-api