Content change due to result of the EU referendum
This is because these are pages that are potentially affected but
the outcome of the EU referendum and have seen a significant spike
in visits since 23 June.

These changes were made by the Content designer via this PR [1].

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.

[1] https://github.com/alphagov/smart-answers/pull/2668