Regenerate regression tests artefacts (ESAP)
This commit shows the state of the regression test artefacts as a
result of the changes to the questions, responses and expected results.

This would also affect other self assessment routes only for the various date entries.

This has lead to the creation of the artefacts within.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.