Use explicit comparison for current tax-year
Using a default value does not make sense from a policy perspective.
If we don’t have a match for one of the selected years it is because
we are not listing them correctly, and something need to be fixed.

This update surfaced an issue with the test as it was assigning a symbol
to the calculator (attribute tax_year), and the application is assigning 
a string.