Add year 2015-16 year to self-assessment-penalties
Regression tests: we have not generated the questions and expected results using
the script: `script/generate-responses-and-expected-results-for-smart-answer`.
The reason is because not all combination of `which_year`, `when_submited` and
`when_paid` make sense. So, to make it compliant with the rest of questions and
expected responses I have manually extended the file:
`test/data/estimate-self-assessment-penalties-responses-and-expected-results.yml`
with the compatible options for year 2015-16