Preparatory Commit for removing 2011-12
This commit shows the preparatory step of removing
2011-12 from the regression test configuration files.

This removal triggers the renaming of 2011-12 to
2012-13 regression files.

This also has similar change for 2012-13 to 2013-14.

For 2014-2015 new file have been created.

This commit doesn't show any changes to the first
question (i.e estimate-self-assessment-penalties/y),

This give assurance that this setup will form the basis
for evaluating the removal of 2011-12 in later commits.

Already existing unit, regression and integration tests passed, giving
assurance that the aforementioned commit/changes have had no
(interfering) effect to other smart answers.