Remove ability to handle multi config set
This commit remove the need of passing the key of a configuration set.

In the past default and future configuration sets were define in the
benefit_cap_calculator.yml file and used to query for data as at when
required.

This removal ensures that only single configuration set can be handle
hence forth.