Update COD slug in marriage-abroad
The slug for Democratic Republic of the Congo has changed into:
democratic-republic-of-the-congo.

Update tests and config files so all code dependant of the old slug is
updated to the new one.

It also include an update to the services file via:

- script/marriage-abroad-services-updates/opposite_sex_cni.yml
- script/marriage-abroad-services-updates/same_sex_removals.yml

As a result, the file `lib/data/marriage_abroad_services.yml` has been
updated.

I have run all the artefact tests and all tests are green.