Refactor consumption of location response
Following the changes from the gds-api-adapter, this commit
changes the way the location response is consumed.

This commit removes the Forwardable and def_delegate method.

Within you will also notice the addition of the
valid_world_location_format? predicate method that takes
a location hash response and decides if it format is as
expected. This method is defined as a class level private
method.

The def_delegate is replaced by the definition and assignment
of the title, details and slug, obtained from the location
hash response passed to the initialiser.

Some cosmetic changes have also been made, mainly renaming
some variables and properties.