Refactor StudentFinanceCalculator
This commit makes changes to the setup of StudentFinanceCalculator.

1. Defining a hash as the only argument

On review of the existing code, I found that the initialise method was defined
and setup with named arguments, also existing (i.e all) instances of StudentFinanceCalculator,
were created with name arguments. By defining a hash, the interpreter is able
convert these name arguments to properties of the defined hash.

2. Define attr_accessor for course_start, :household_income, :residence, :course_type

This commit also allow for modification of the instance variables belonging to
StudentFinanceCalculator, this is neccessary and will help gather user response as the
use progress through this smart answer.