Instantiate calculator to calculate grant and loan
This commit reverts back to the former way that sets up the calculator,
within the whats_your_household_income? question before calculating
maintenance_grant_amount and maintenance_loan_amount.

This help avoid the problem where the smart-answers state calculator object
that was previously set in the first question.

I tried using the on_response block like this.

`
  on_response do |response|
    calculator.household_income = response
  end
`

`
  on_response do |response|
    calculator.course_start = start_date
    calculator.residence = where_living
    calculator.household_income = response
  end
`

I also tried moving grant and loan blocks to next question

i.e the following blocks the next question:

- maintenance_grant_amount
- maintenance_loan_amount

In my opinion this was suppose to allow the state calculator object time to
properly set and hold the users household income input.

The aforementioned approach, appears to work fine for calculate blocks like
tuition_fee_maximum as they depend on already define and set attributes
(i.e TUITION_FEE_MAXIMUM).

But in the case of maintenance_loan_amount, there appears to be more to it
that means the eyes.

After trying out the approaches mentioned above, the symptoms still persist.

I discussed this with the team we decided to move ahead with this, to
sort it out later, mainly because of the limited time available.

[1] https://github.com/alphagov/smart-answers/blob/master/lib/smart_answer_flows/student-finance-calculator.rb