Add outcome to render marriage abroad for any country
We are exploring inlining content for marriage abroad so content designers
have less issues with BAUs. By inlining content, we mean creating a unique
govspeak partial per outcome with no content fragmentation.
We can take this approach because all the outcomes share the same set of
questions to reach an outcome.

So this commit includes:

* A Generic outcome that renders the content of any country. It leverages
the function described below.
* Generic function to return the path to an outcome `path_to_outcome`

There are going to be 18 possible outcomes for a country. The function
determine the path to an outcome returning an array with the different
options for that path.