Refactor Getting a statutory declaration section
The Getting a statutory declaration text is used 
in 6 of the 12 outcomes for Italy.

I have taken the 4 most common instances (where
the partner is not british) and moved them into
a shared partial template.

The original partial template has been removed as
it is no longer needed.

No test artefacts were affected by this change.