Add hints partial template for BCC
This commit adds a hints partial template to the benefit cap calculator.

This helps reduce the duplicate hint content that has been identified across
several questions.

The replace of duplicate the content with the hints partial has not caused any
regression test artefacts to be regenerate or deleted. Only checksums
are affected due to change in the templates content