Use content store for breadcrumbs, sidebar & metadata
This commit changes the way we render breadcrumbs, sidebar and
content-store. It is part of similar work in all "mainstream"-like
applications on GOV.UK (https://trello.com/c/mOeDK914).

Instead of the content-api, we now use the content-store as the data
source for everything. We also use the `govuk_navigation` gem to
generate GOV.UK component data.

https://trello.com/c/49441pPb