Preparatory: Add Austria to which_country list
This is a preparatory commit adding Austria to the which_country response
list.

This is in preparation for the content changes to be made in later commit
where a death happened in Austria and the user lives in North Korea

This commit also shows the affected/changes to regression test artefacts
as a result of the changes made to add new countries.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.