Implement TODO on currently_in_north_korea? method
This commit adds the required predicate check to currently_in_north_korea?
calculator method.

It appears had been omitted in this commit [1], based on findings the
next_node_calculation block due the absence of the response argument,
So the response within it returned nil. Probably some investigation or refactoring
exercise was carried out and some how the TODO comment mistakenly omitted.

This appears for be responsible for the rendering of the oru_result outcome and
not the expected north_korea_result.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

With reference to the Trello card [2], this clearly isn't only affecting Austria only.

Within you will notice that the regression test artefacts also other countries

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.

[1] https://github.com/alphagov/smart-answers/commit/e3d5220a0f643b9bf7ba7190378360b8c43f26b6
[2] https://trello.com/c/gpJqutRR/208-possible-bug-in-register-a-death-north-korea