Use fees table partial.
Marriage abroad fees have already been set up as a RatesQuery. It makes
it easier to update the fees if we reuse the consular fees partial that
has already been created to insert them.

Extract the payment method conditional into a partial template so that
it can be used in the Italy specific templates and the generic
services and fees partial.