Remove user’s previous birthday from outcome.
The calculator tells you how much you pay to top up your State Pension
based on:

* the age you were on the birthday before last
* the age you were at your last birthday (ie your current age)
* the age you'll be on your next birthday - this only displays
if you reach your next birthday by 5 April 2017, when the top-up scheme
ends.

The first point is causing users some confusion. As, even though the
information is correct, they can’t act on it as that year (the birthday
before last) is in the past.

Change the calculator so that, instead, it only shows how much you pay
to top up your State Pension based on:

* your current age
* the age you'll be on your next birthday if your birthday is before
5 April 2017

Update unit and integration tests with the new expected behaviour.

Update the default date used but state-pension-topup in the regression
tests so that the test artefacts also reflect the new expected
behaviour.