Create a template for marriage-abroad title in Italy
There is no single rule to infer the content title for a particular
country in marriage-abroad.

Rather than have a lot of logic in the main country template that adds
exceptions for some countries, move this logic into a specific title
partial template for a country.

While this approach will create duplicate title partials, it will make it
easier for content designers to update the title for one country without
affecting the others.

Update `MethodMissingHelper`

Following the falling test [1]  and because it is used to create a snapshot of
the content that will be registered. We are forcing the `calculator.ceremony_country`
to be Italy

[1] test/integration/indexable_content_test.rb

No test artefacts were affected by this refactor.