Use services&fees partials: marriage-abroad Sweden
The content has been flattened but we want to keep all the work that
alreadys exists with the consular services and fees. For this reason we are
using the existing partials to render this information.

All regression tests are passing, so we are confident that existing
content has not been changed.