Content changes for Kenya (UK Resident)
This commit updates content changes where the user is a British
citizen and living in the UK.

This commit also shows the regression test artefacts as a result
of the content changes (i.e content changes).

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.