Change content in all same sex outcomes Italy
The way we structure the marriage-abroad calculator for Italy is
changing.

Rather than 4 questions that lead to 18 possible outcomes, we are
only going to ask users:

1) Where do you want to get married?
2) Is your partner of the opposite sex, or the same sex?

These questions are questions 1 and 4 for the other countries.

This means that rather than 18 possible outcomes for Italy, there will
only be two: opposite-sex and same-sex.

To prepare for the Italy flow changes, I have changed all of the
same-sex outcomes to have the new content, and regenerated the
test artefacts.

These changes form a benchmark, and going forward, the test artefacts
for same-sex marriages should not change.

For ease I am hard-coding the fee amounts at this time. After the
same-sex outcomes have been consolidated into one template I will
get the fees from the rates file.