Change flow for Italy in marriage-abroad
Change the flow for Italy so that it only asks the user two questions:

1) Where do you want to get married?
2) Is your partner of the opposite sex, or the same sex?

Update the integration tests, so that only two responses are collected.

Regenerate the expected results file to remove questions 2 and 3.

Regenerate the test artefacts. As expected, the path to the test
artefacts has changed even though the content has not. Rather than 18
artefacts for Italy, there are now only 2.