Change default time for marriage-abroad in tests.
The default time that the regression tests uses is “2015-01-01”. This
causes problems when you want to check that the rates used in an 
outcome is correct.

In this story, I will be using the fees from the rates file. I need to 
be sure that when restructure the smart-answer that I am picking up the
current fees for Italy.

This commit, updates the fees in the test artefacts for all countries to 
use the latest amount.