Remove unused template files for Italy
The outcomes for Italy have been consolidated into two templates.
The templates created to inline the content for Italy are no longer being used
and can be removed.

Change the path_to_outcome for Italy in MethodMissingHelper

This fixes the indexable content error [1]

As the default template location for path_to_outcome is removed in this commit,
the MethodMissingHelper has been updated to a location that does exist.

No test artefacts were affected by this change.

All unit and integration tests are still passing.

[1] https://ci.dev.publishing.service.gov.uk/job/govuk_smartanswers_branches/6990/console