How to generate checksums example
Add examples of how to (re-)generate checksums.

When I tried to run the script, I didn't know I had to use the rails
command or what parameters looked like. After being shown, I have added
examples to the docs.

The first example is how we would normally run the script. The second is
if you want to ensure that a directory of files changed for a flow are picked
up to generate checksums for. A third example usage specifies the path
to a single file.

The last two may not be necessary, but I have documented an example
regardless as the script supports these features.