Route paths for Mozambique to generic outcome
Send all paths through marriage-abroad from Mozambique to the generic countries outcome.

Add a function that decides whether a country should be routed to
outcome_marriage_abroad_in_country. Having a function means that we
don’t need to update the flow every time we had a country that has an
outcome per path. In future PRs I might consider moving the list of
countries to a global list.

Recreate the expected responses and results file to point to the generic
country outcome

Govspeak is registered as a MimeType so that the templates can be
found and rendered.

No test artefacts were affected by this change.