Change date made redundant in `employee` calculator
Change the expected responses in calculate-employee-redundancy-pay to
set the date made redundant to 2013.

The content does not change, only the path to the questions and
outcomes. The responses have been changed because 2012 will no longer be a
valid redundancy date in the calculator.

The subsequent changes should not generate any changes to the test artefacts.