Add next year to the redundancy date selector
The first question, what is your redundancy date, limited users to adding a
date that is in the current year (e.g. 2016). If you know that you are being
made redundant in the new year, you need to be able to check a date in
the following year.

No test artefacts were affected by this change.