Automatically determine latest start date
Automatically determine the latest year that a user can select in the
`What date was your employee made redundant?` question in both redundancy pay
calculators.

If the current date is in the beginning of the year (between January and
August), the user should only be able to select a redundancy date that
is before the end of the current year.

If the current date is in the later part of the year (between September
and December), the user should be able to select a redundancy date that
is before the end of the following year.

Affected calculators:

* calculate-your-redundancy-pay
* calculate-employee-redundancy-pay

No test artefacts were affected by this change.