Automatically determine earliest start date
Automatically determine the earliest year that a user can select in the
`What date was your employee made redundant?` question in both redundancy pay
calculators.

If the current date is in the beginning of the year (between January and
August), the user should only be able to select a redundancy date that
goes back as far as the beginning of the year, 4 years ago.

i.e if the date is 01/08/2016, the earliest date a user can select is
01/01/2012.

If the current date is in the later part of the year (between September
and December), the user should be able to select a redundancy date that
goes back as far as the beginning of the year, 3 years ago.

i.e if the date is 01/10/2016, the earliest date a user can select is
01/01/2013.

Affected calculators:

* calculate-your-redundancy-pay
* calculate-employee-redundancy-pay

No test artefacts were affected by this change.