Regenerate test artefacts with latest fees
The current time that the regression tests use to select rates is in 2015. Any
rate changes after that date are not reflected in the test artefacts.

A later part of this story involves changing the questions asked to users who are
getting married in France, eventhough the final content they are presented with is
not changing at the moment.

update the date that the marriage abroad regression tests use to get the most
current fee rate, so that it is easier for factcheckers to review these changes.