Change France flow to use new templates
Change the flow for France to use the new templates created in the previous commit.

This involves a few steps:

1. Add an exception to the `path_to_outcome` to route the flow
for France to the new outcome templates. This exception is required because the
flow for France only asks two questions, so France only has two outcomes. This
doesn't fit the current pattern for `path_to_outcome` which assumes that the user
has answered four questions.

2. Make the marriage_or_pacs attribute in the calculator accessible. This
is so that it can be used to by the path_to_outcome method to route the flow
for France to the new outcome templates.

No test artefacts were affected by this change.