Create an outcome per path for the Philippines
The goal of this story is to change the flow for the Philippines to only ask
one additional question:

Is your partner of the same or opposite sex?

And for each of the paths generated by that question to have it's own outcome.

The approach I've decided to take is to:

1) Make an outcome for each path as it currently exists (this commit)
2) Change the current flow to use these templates
3) Modify the templates, so that all of the opposite sex outcomes have the same
content and all of the same sex outcomes have the same content.
4) Change the flow to only ask two questions, and consolidate the outcome
templates so that there are only two, one for opposite sex and one for same sex
5) Clean up the Philippines from the rest of marriage-abroad (templates, flow
and lists)

No test artefacts were affected by this change.