Make all of the same sex outcomes the same.
Make the content in all of the same sex outcomes the same. I have
arbitrarily chosen ceremony_country/partner_british as the outcome
whose content has been copied to all of the other same sex outcomes.

Making this change and now means that we can see the update to the test
artefacts in isolation from future updates to the flow.