Change Philippines flow to only ask two questions
Change the flow for Philippines in marriage-abroad to only ask two
questions:

1) Where are you?
2) Is your partner the same or opposite sex?

Update the integration test to only add two responses for the
Philippines.

Update the expected responses file for the Philippines.

The 18 old test artefacts have been deleted and replaced with 2 new
ones, though the content of the test artefacts has not changed.