Change to start page body content.
According to the this document [1] this change has been made because
there is no need for ‘now’ since this is the current state of
affairs and it doesn’t need to be presented in context of
what came before. And to remove Oxford comma.

This commit shows the affected/changes to checksums and regression test
artefacts as a result of the content change.

I have reviewed the changes to the regression test artefacts and it
shows the difference as a result of the aforementioned change.

At this point existing unit, regression and integration tests passed,
giving assurance that the changes have had no (interfering) affect to
other smart answers.

[1] https://docs.google.com/document/d/1vyWhULg4cVGd4DJnhfBsbEtcOz077VeozWKSSdu0kSg/edit