Copy changes for opposite sex outcomes in uk
Copy changes for all opposite sex outcomes for Latvia where the user is 
in the UK.

Added `affidavit_or_affirmation_for_marriage` to services data for 
Latvia as the new content expects a fees table with this fee in it.

There is a discrepancy between the method of payment in the google doc
and what is shown in the outcome. That is because the payment partial
for all Latvian consular fees is common for all outcomes and has already
been set to `pay_in_local_currency_ceremony_country_name`

Test artefacts have been modified as expected.