Preparatory: Update dates to 2017/2018
In preparation for the 6th April rates update, this commit replaces
2012/2013 dates with 2017/2018 dates. This is a straight swap & triggers
this generation of new regression test artefacts for2017/2018 and removal
of 2012/2013.

Before this I tried using both sets of dates but the tests took
ages to run. I decided to make this swap based on test run performance.

I recognised that it appears somewhat confusing, the intent is to make clear
the rate changes that will come after this. These sets the basis for evaluating
the content changes in the commits to come, as the configuration.yml entry for
calculate-statutory-sick-pay has been updated to a future date.

Based on the above, I have reviewed the changes to the regression test
artefacts and they appear as expected.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no (interfering) effect to
other smart answers.