Add implementation to remove existing options
With reference to this commit [2], adding the ability to multiple choice to remove
option as at when necessary, help avoid the problems for accumulating options as
a user browse through different smart answers flows.

In this case, the student finance forms has changes for uk-full-time that include
showing only 2017/18 and 2016/17 options whilst keeping the existing 2015/16
and 2016/17 options.

To avoid duplicate questions an initial attempt was made with a pre-calculate [3].
This provides the much needed access to the state (scope) object [4] , in this case
type_of_student.

This attempt resulted in duplication of options. hence why the need to reset add add
options is required.

[1] https://github.com/alphagov/smart-answers/blob/master/doc/storing-data.md
[2] 6f62dd6296ba7155f55d7d83d760f5f2eb5aee37
[3] https://github.com/alphagov/smart-answers/blob/master/doc/flow-definition.md#precalculatevariable_name-block
[4] https://github.com/alphagov/smart-answers/blob/master/doc/flow-definition.md#scope