Remove Question one (SEC)
According to [1], the Content team have requested the removal of
question 1 to be replaced by question 2.

The reason for this , is because for capital allowance a new or existing
business is no longer a criteria.

This has been removed and the questions sequence reordered. See [2]

NB: Renumbering questions:

This is a preparatory commit for subsequent changes to fit in more
naturally and incrementally. As part of this, I have added a comment to
act as a “marker” for where a new Question 4 which we have not yet
implemented but we will get slotted in in later commits.

This is so that we can plot an overview of the changes that the
Content Team have requested.

I have included the regression test artefacts within to show areas
affected by these changes

Based on the above, I have reviewed the changes to the regression
test artefacts and they appear as expected.

Already existing unit, regression and integration tests passed,
giving assurance that the aforementioned commit/changes have had
no (interfering) effect to other smart answers.

[1] https://docs.google.com/document/d/1vyWhULg4cVGd4DJnhfBsbEtcOz077VeozWKSSdu0kSg/edit
[2] https://docs.google.com/document/d/1XkyRR2iql7I6kRyXCOCpxGtmIdOnzAfsHltnyTbQiSk/edit