Add notice to sign
It is a requirement for passports to be signed before they can be used.
This change is to add a notice to make clear that overseas recipients of
new British passports should sign.

Test artefacts of regression tests on this Smart Answer are affected
and updated according to the expected change. No other tests are impacted
by this minor change.