Update oru_result.govspeak.erb
1. Changed second bullet from: 
'the child’s mother is married to a man who is not a UK citizen, even if the child’s father is a UK citizen'
To:  'at the time of the child’s birth the child’s non-British mother was married to a man who is not a UK citizen, even if the child’s biological father is a UK citizen'
Reason - child is given their citizenship status by their mother's  spouse (rather than biological father).  
2. Changed lead-in sentence from: 
'You won’t be able to register a birth if with the UK authorities if any the following apply:'
To:  'You won’t be able to register a birth with the UK authorities if any the following apply:'
Reason: Typo
3.  Removed: 'You can also register the birth with the UK authorities.'
Reason: We already say:  'You won’t be able to register a birth with the UK authorities'.