Update current time for state-pension-popup tests
This pull request is trying to replicate and fix a bug that occurs if
your birthday falls in the last 3 months of the scheme. 

Updating the current time for the regression tests allows us to baseline
regression test content with what is currently on the live system.

Future commits should see this content change.