Only show one age, if no more birthdays before end
The state-pension-topup calculator should only show one topup age if the user
doesn't have any more birthdays before the scheme ends.

The original loop to get the number of birthdays, works out how old you are going
to turn in a calendar year.

For the final year of the scheme, this needs to be amended to be how old you are
going to turn in the first 3 months of the year, and if you have a birthday in
the first 3 months, it should take into account whether your birthday has already
passed or not.

That wasn't happening with the previous hack. If your birthday is between 1st
January and 5th April it was showing two ages, regardless of whether you'd already
celebrated your birthday.