Refactoring the calculate blocks
This commit attempts to identify and move code from the calculate blocks
in the simplified expense checker tool.

This has done in a test driven was with unit test intoduced to interpret the
behaiviour of the various bits of code moved.

The exist integration test act as validation that the code moved has had not impact to the flow of this smart answer.

NO REGRESSION TEST ARTEFACTS HAVE BEEN AFFECTED