Add an option for non-EEA family member tenant
Update the `what_nationality?` question to allow users to select that
the tenant is a non-EEA family member of a EU, EEA or Swiss citizen.

If the user selects the `non-era` option, currently they are sent to 
`outcome_can_rent`. This is a place holder until the new question they
should be routed to is implemented in the next commit.

Unit tests have been updated to accept the `non-eea` option.
Regression tests have been updated to test this new option.