Add permanent residence card question
If the tenant is a non-EEA resident, ask if they have permanent
residence card.

This question is multiple choice, but both responses route to
`outcome_can_rent`. This is because the question that `no` should route
will not be implemented until the next commit.

Unit and regression tests have been updated to include the new question.