Re-implement question 13, has residence card
Question 13, `has_residence_card…` has been re-implemented. This is where
the user should be routed to if they answered no to the previous question,
the new question 5, that asks if the user has a permanent residence
card.

This question is multiple choice, but both options go to same outcome.
That is because the outcome and question they should route to haven’t
been implemented yet.

Unit and regression tests have been updated to include this question.