Reimplement question 9, has_documents?
If the user answered “no” to the question implemented in the previous
commit, `has_residence_card_or_eu_eea_swiss_family_member?`, they should
be directed to `has_documents?`. This question has been previously
implemented and commented out.

This is a yes/no question. Currently both options direct the user to
`outcome_can_rent`. This is because the route for `no` hasn’t been
implemented yet.