Reimplement question 12, time limited leave to remain
If the user answered `no` to question `has_documents?` they should be
directed to question `time_limited_to_remain?`. This question had been
previously commented out, but has been reimplemented here.

This is a yes/no question, but both responses point to the same outcome.
This is because the question that answering `no` should direct the user
to hasn’t been implemented yet.

Unit and regression tests have been updated to include this question.