Implement final question in non-EEA flow
If the user answered `no` to question 14 (`waiting_for_documents?`) they
should be routed to the `immigration_application?` question. This question was
previously commented out but has been re-added to the flow.

The outcomes routed to from the `yes` and `no` responses in the question
have already implemented and are reused here.

Unit and regression tests have been updated to include this question.