Fix 2016-2017 tax year for mat_pay partial
This is a preparatory commit to make for the changes for the 2017-2018
financial year.

In the `_mat_pay` outcome, the 2016-2017 is not specifically
referenced. It is just referred to as the current tax year. Of course,
when we get to April 6th 2017, this will no longer be true. But if the
due date of the child is in the 2016-2017 tax, the user would still
see this text.

I have added a calculator to check whether the due date plus 39 weeks
falls in the 2016-2017 financial year. This calculator is being used in
the outcome, and I’ve styled the text to match the other years.

There is a bit of oddness in the test artefacts.
For due dates in 2012: We don’t have the rate for this year. As doesn’t
fall in any other financial year, it was picking up the text and rates
for 2016-2017, but obviously this wasn’t correct.

Artefacts with an extra line: If the due date plus 39 weeks falls across
two financial years, only one line was being displayed. That means that
users were not being alerted to the fact that they could receive different
pay rates at different times of the 39 weeks.