Fix 2016-2017 tax year for additional_pat_pay The `additional_pat_pay` partial outcome template displays text relating to the 2015-20176 tax year if the pay date range falls in the 2015-2016 tax year. However this text is also displayed if the pay date range doesn't fall in the 2013-2014, 2014-2015, 2015-2016 tax years. This means that the text could be displayed for the 2016-2017 and 2017-2018 tax years.
Also, I've noticed that there is not text referencing the 2016-2017 tax year.
I believe this to be an error, and have fixed both situations by re-using the
2016-2017 calculator function and updating the text.

There is a bit of oddness in the test artefacts.
For due dates in 2012: We don’t have the rate for this year. As doesn’t
fall in any other financial year, it was picking up the text and rates
for 2015-2016, but obviously this wasn’t correct.

Artefacts with an extra line: If the due date plus 39 weeks falls across
two financial years, only one line was being displayed. That means that
users were not being alerted to the fact that they could receive different
pay rates at different times of the 39 weeks.