Preparatory: Increase earnings to 28001
This replaces the the earnings above the personal allowance for the
current tax year, replacing 27701 with 28001.

Also the regression tests configuration entry for
calculate-married-couples-allowance smart answer has been updated to a
future date within the next financial year.

I am changing the married couple personal earnings from 27701 to 28001.

In the current tax year (i.e 2016-2017), the personal allowance for a
married couple is 27700, but this has now been increased  to 28000 in
the next year (i.e 2017-2018).

These changes here have been made with the intention of making changes
in the next commit to the earnings amount. The changes should only
affect lines with earnings related content.

This commit only changes the file names of the test artefacts, and the
changes to the amounts should only occur in the next commit.

However, there exist some artefacts that have not only been renamed but
also changes to content. It is not entirely clear why this has occurred.

It appears these two scenarios concern those not paying into a pension,
hence the pension contribution are set to zero.

Further information about these changes can be found in [1] and [2]

I have included the regression test artefacts within to show areas
affected by these changes

Based on the above, I have reviewed the changes to the regression
test artefacts and they appear as expected.

Already existing unit, regression and integration tests passed,
giving assurance that the aforementioned commit/changes have had
no (interfering) effect to other smart answers.

[1] https://docs.google.com/document/d/1AhqvcSM99gGdCeeFh6LfxBDfcOPq_j11prvJuvJqVGs/edit#
[2] https://github.com/alphagov/smart-answers/blob/master/lib/smart_answer_flows/calculate-married-couples-allowance.rb#L70