Fix 2016-2017 tax year for mat and pat shared pay
The `mat_shared_pay` and `pat_shared_pay` partial outcome templates
display text relating to the 2015-2016 tax year if the pay date range
falls in the 2015-2016 tax year.
However this text is also displayed if the pay date range doesn't fall
in the 2013-2014, 2014-2015, 2015-2016 tax years. This means that the
text could be displayed for the 2016-2017 and 2017-2018 tax years.

Also, I've noticed that there is not text referencing the 2016-2017 tax
year. I believe this to be an error, and have fixed both situations by
re-using the 2016-2017 calculator function and updating the text.

Artefacts with an extra line: If the due date plus 39 weeks falls across
two financial years, only one line was being displayed. That means that
users were not being alerted to the fact that they could receive
different pay rates at different times of the 39 weeks.