Add 2017-2018 tax year to lower_earnings_amount
The `lower_earnings_amount` is changing in the 2017-2018 tax year.
Updated the `lower_earnings_amount` calculator function to add this
value and updated the default value to match.

Unit tests have been updated to include the 2017-2018 financial year.

The regression tests for `2012` have been updated because it picks up
the default value and that value has changed.