Add new question what_sort_of_passport?
If the user selects Estonia in the first question, they are now being
redirected to a new question.

This new question currently directs all users to outcome 1, the same
outcome that Estonian passport holders were directed to before this
question was added. That is because the new question will be fully
implemented in the next commit.

A calculator function and unit tests have been added to check whether
Estonia has been selected.

Integration tests have been updated to include the new flow for Estonia

Test artefacts have been updated as expected.