Create new study waiver outcome for Taiwan
The routing for Taiwan seems to have gone a bit wrong. All outcomes for
Taiwan were the user is eligible for a waiver point to text about 
applying for a study visa. This is obviously wrong, because sometimes
the user needs to apply for a medical visa or a standard visitor visa.

I’ve decided to fix this bey creating a different template for Taiwan
for each type of visa.

In this commit, I have created a new template for the study waiver.
The test artefacts haven’t been affected because the text in the new
outcome matches the text in `outcome_taiwan_exception`. This won’t be
the case in the following commits.

Integration tests have been updated.