Preparatory commit: Employee redundancy pay
Add test artefacts for the next tax year to
`calculate-employee-redundancy-pay`.

The statutory redundancy rates are changing. To be able to easily see
the effect of the rate changes, two preparatory steps were added:

1. Add a redundancy date in the next tax year to the test artefacts

2. Set the `current_time` for the test artefacts for to a date in the
next tax year.

This means that when the test artefacts are created it picks up the
most recent redundancy rates from the rates file (for 2016-2017) but
when the new rates are added in a later commit, these files will change.