Remove unused USA code from marriage-abroad
The USA has it’s own outcomes, so logic specific to the USA can be
removed from other outcome templates and lists.

The unit tests have been updated because we are not using the
`country_name_lower_case_prefix` in the USA outcomes, so we do not need
to test for it.

No test artefacts were updated by this change.