Fix the total claim content
Originally question one asked about vehicle and business locations,
so the content in the template was only being rendered if the user only
selected live_on_business_premises. Now that question one has been split,
list_of_expenses will always contains a value for vehicle even if the value
is no vehicle so the conditional in the template would never evaluate to true.

So now we changed the contional to match the intention of the orginal, which is
that the user only selects only live_on_business_premises and has no vehicle.

The regression tests artefacts are now as expected.