Add un-publish rake task and implementation
This commit adds implementation for un-publishing a smart answer from the
publishing-api.

This uses the publishing-api to un-publish the given smart answer.
A smart answer's content-id  is required to be passed to the rake task.

(i.e `rake publishing_api:unpublish[content-id]`)

When this rake task is invoked a post request [1] is composed and made.
This transitions an edition of a document into an unpublished state. The
edition will be updated or removed from the live content store and sets
it to type of gone with status 410.

Tests have been added that describe this implementation.

[1] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#post-v2contentcontent_idunpublish