Redirect smart answer paths to new destination
This commit adds implementation that redirects a path to new destination.

This has become necessary as part of the process of retiring a smart answer.

Once a smart answer has been unpublished it paths may need to point to a new
destination.

This implementation provides a rake task level interface that helps
create a redirect format with a prefix type on the content store.

This uses the publishing and router apis to establish this the redirect between
path and it's new destination.

(i.e `rake publishing_api:redirect_smart_answer[path,destination]`)

When this rake task is invoked requests (i.e [1] and [2] from [3]) are composed
and made. This request publishes the redirect with the supplied parameters.

Tests have been added that describe this implementation.

[1] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#put-v2contentcontent_id
[2] https://github.com/alphagov/publishing-api/blob/master/doc/api.md#post-v2contentcontent_idpublish
[3] https://github.com/alphagov/content-store/blob/master/doc/redirect_item.md